Revert "Merge branch 'main' into dev-noah"
This reverts commit 74fc713b618a79b132a7cc521a990e3ae9d233b1, reversing
changes made to 2a87a4aabe117c09172ee525099bb401f34eafc0.
