+++ lib/game_route.dart
@@ -7,11 +7,9 @@ import 'package:provider/provider.dart';
 WebViewEnvironment? webViewEnvironment;
 
 class GameRoute extends StatelessWidget {
-  final String gameUrl;
-  final String gameShareableUrl;
-  GameRoute({required this.gameUrl, required this.gameShareableUrl});
   @override
   Widget build(BuildContext context) {
+    var appState = context.watch<MyAppState>();
     final theme = Theme.of(context);
     final bool isDarkMode = theme.brightness == Brightness.dark;
     return Scaffold(
@@ -25,7 +23,7 @@ class GameRoute extends StatelessWidget {
                     isInspectable: kDebugMode,
                   ),
                   initialUrlRequest:
-                      URLRequest(url: WebUri(gameUrl ?? "")),
+                      URLRequest(url: WebUri(appState.gameUrl ?? "")),
                   onWebViewCreated: (controller) {},
                   onLoadStart: (controller, url) {},
                   onLoadStop: (controller, url) async {
