+++ pubspec.yaml
@@ -1,7 +1,7 @@
 name: dailytrojan
-description: "The Daily Trojan Mobile App."
+description: "A new Flutter project."
 publish_to: 'none'
-version: 0.0.2+2
+version: 0.0.1+1
 
 environment:
   sdk: '>=3.4.3 <4.0.0'
